tests/ports: fix build failure on x86_32
pagesize is 64bits on Linux/x32 in case we load a 64bits sysfs dump.
tests/ports only run on Linux, so change pagesize to uint64_t here too.

+++ tests/hwloc/ports/include/freebsd/sys/mman.h
@@ -7,6 +7,7 @@
 #ifndef HWLOC_PORT_FREEBSD_SYS_MMAN_H
 #define HWLOC_PORT_FREEBSD_SYS_MMAN_H
 
-extern int getpagesizes(size_t pagesize[], int nelem);
+/* use uint64_t for pagesize since this port test runs on Linux */
+extern int getpagesizes(uint64_t pagesize[], int nelem);
 
 #endif /* HWLOC_PORT_FREEBSD_SYS_MMAN_H */
